Double check that the export file has been uploaded correctly.
In WordPress 3.3 and 3.3.1 it is possible for wp_handle_upload() failure caused
by bad permissions on the upload directory to be report incorrectly. Instead of
an appropriate error message the user is told that they uploaded an invalid WXR
file with a missing or incorrect WXR version number. See #WP19800, #WP19720.


git-svn-id: http://plugins.svn.wordpress.org/wordpress-importer/trunk@504381 b8457f37-d9ea-0310-8a92-e5e31aec5664